api/requirements.txt
fastapi==0.63.0
firebase_admin==5.1.0
google-cloud-logging==2.2.0
google-cloud-pubsub==2.3.0
python-dotenv==0.15.0
pytz==2021.3
requests==2.25.1
starlette-context==0.3.3
uvicorn==0.13.4